Simple demo for making, storing and playing an experiment
