+++ Assets/Settings.cs
@@ -1,7 +1,10 @@
 ﻿using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 
 public class Settings : MonoBehaviour
 {
-    public static int environment;
+    public static int index = 0;
+
+    public static List<EnvironmentConfiguration> environments; 
 }
